+++ private/planner/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to the Planner Pro application will be documented in this fi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.2] - 2026-01-02
+
+### Fixed
+- Fixed 412 Precondition Failed errors when creating goals by fetching details first to get proper etag
+- Fixed goal target date not saving when editing (now properly converts to/from ISO format)
+- BUCKET_GOAL_MAPPINGS task now named "[System] Bucket-Goal Mappings" and marked complete to be less visible
+- System tasks filtered out of goals list to prevent confusion
+
 ## [3.2.1] - 2026-01-02
 
 ### Changed